Converted lots of TensorGraph layers to Keras layers
